Use `cfg(target_has_atomic)` to conditionally compile impls
This makes this crate a lot more useful to targets where not all atomic
sizes were supported. Previously, this crate would simply not compile
on those targets, because (e.g.) `AtomicU64` was not found. With this
commit, `u64` just does not implement `Atom` or `PrimitiveAtom` on those
targets. (And neither does `f64` implement `Atom` anymore.)

This is not yet an optimal solution, as platform support for atomics
are not just differentiated by size, but also by whether they support
some features. Sadly, for those, there are no `cfg` flags available
on stable yet.

This commit raises the MSRV to 1.60