Add `Send + Sync + Unpin` bound to `PrimitiveAtom::Impl`
This is not a breaking change as the trait is sealed. The `Impl`
associated type was `doc(hidden)` before as no one should care about it.
That changed when we needed a `const` way to create `Atomic`. This
change is useful as the whether `Atomic` implements `Send`, `Sync` and
`Unpin` currently depends on the `Impl` type. That's inconvenient.
`Atomic` should always implement `Send`, `Sync` and `Unpin`, so the
bound was added.